% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essc.R
\name{essc}
\alias{essc}
\title{essc}
\usage{
essc(Adj.Matrix, alpha, Null = c("Binomial", "Poisson"),
  Num.Samples = nrow(Adj.Matrix))
}
\arguments{
\item{Adj.Matrix}{Adjacency matrix of the network for which you'd like to find communities}

\item{alpha}{False discovery rate. Value between 0 and 1 that sets an upper limit on the false discovery rate of the hypothesis testing procedure in ESSC}

\item{Null}{The null distribution used for comparing the observed number of connections between a single vertex and a community of vertices. Can be set to either "Binomial" or "Poisson." Default is "Binomial"}

\item{Num.Samples}{The number of randomly selected neighborhoods used to initiate the ESSC algorithm. The maximum number is the number of vertices in Adj.Matrix. Defaults to the number of vertices in Adj.Matrix}
}
\value{
a list containing the objects 
\itemize{
   \item Communities: a list of identified communities
   \item Background: a numeric vector identifying vertices that did not belong to any statistically significant community
}
}
\description{
Identify statistically significant communities in undirected networks
}
\examples{
net <- stochastic.block(n = 1000, k = 3, P = cbind(c(0.1, 0.01, 0.01), c(0.01, 0.1, 0.01), c(0.01, 0.01, 0.1)), sizes = c(300, 300, 400))
results <- essc(net$Adjacency, alpha = 0.10, Null = "Poisson")
print(results$Communities)
}
\author{
James D. Wilson
}
\references{
\itemize{
    \item Wilson, James D.,Wang, Simi, Mucha, Peter J., Bhamidi, Shankar, and Nobel, Andrew B. (2014). “A testing based extraction algorithm for identifying significant communities in networks.” 
    The Annals of Applied Statistics Vol. 8, No. 3, 1853-1891 
}
}
\keyword{community}
\keyword{detection,}
\keyword{extraction}

