% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FB_edgelist}
\alias{FB_edgelist}
\title{A personal Facebook edgelist displaying the relationship between two nodes}
\format{A data frame with 16750 rows and 3 variables:
\describe{
  \item{Vertex.1}{the vertex of one node, undirected}
  \item{Vertex.2}{the vertex of another node, undirected}
  \item{Number.edges}{The number of edges from one edge to another}
}}
\usage{
FB_edgelist
}
\description{
A dataset containing the vertices and edges between two nodes
}
\keyword{datasets}
