% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FB_labels}
\alias{FB_labels}
\title{The labels indicating the type of association}
\format{A data frame with 561 rows and 2 variables:
\describe{
  \item{Individual}{the individual from Facebook}
  \item{Location}{what association the individual has with James.
  1 meaning an associate from work. 2 meaning an associate from college. etc}
}}
\usage{
FB_labels
}
\description{
A dataset containing the type of relationship between an individual and James Wilson
}
\keyword{datasets}
