% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_block_model.R
\name{stochastic.block}
\alias{stochastic.block}
\title{stochastic.block}
\usage{
stochastic.block(n, k = 2, P, sizes, random.community.assignment = c(FALSE,
  TRUE), vector.output = c(FALSE, TRUE))
}
\arguments{
\item{n}{Number of nodes}

\item{k}{Number of blocks. Default is 2}

\item{P}{k x k matrix where the (i,j)th entry specifies the probability of connection between community i and j}

\item{sizes}{an array of length k where the ith entry specifies the size of the ith block. This must sum to n}

\item{random.community.assignment}{logical specifying whether or not the labels are assigned randomly. If FALSE, labels are assigned in order starting with the upper left corner of the matrix. Default is FALSE}

\item{vector}{logical specifying whether or not the returned adjacency matrix should be returned as a vector of edge variables. Default is FALSE}
}
\value{
a list containing the objects 
\itemize{
   \item Adjacency: either an adjacency matrix of the generated realization, or if vector == TRUE, an edgelist of the generated realization
   \item Membership: a numeric vector specifying the community label of each vertex
}
}
\description{
Generate a realization of the stochastic block model with non-overlapping blocks
}
\examples{
net <- stochastic.block(n = 300, k = 2, P = cbind(c(0.1, 0.01), c(0.01, 0.1)), sizes = c(100, 200))
image(net$Adjacency)
}
\author{
James D. Wilson
}
\keyword{SBM}
\keyword{block}
\keyword{community}
\keyword{detection,}
\keyword{model,}

